feat: dynamic grid component
