+++ components/items/ItemsGrid/ItemsGrid.vue
@@ -3,11 +3,11 @@
     <!-- TODO: breadcrumbs filter here -->
     <p>total: {{ total }}</p>
     <hr />
-    <div ref="container" :style="gridCols">
+    <DynamicGrid>
       <div v-for="(nft, index) in nfts" :key="`${nft.id}=${index}`">
         <ItemsGridImage :nft="nft" />
       </div>
-    </div>
+    </DynamicGrid>
     <div ref="reachBottom">bottom</div>
   </div>
 </template>
@@ -15,6 +15,7 @@
 <script setup lang="ts">
 import { useIntersectionObserver } from '@vueuse/core'
 
+import DynamicGrid from '@/components/shared/DynamicGrid.vue'
 import ItemsGridImage from './ItemsGridImage.vue'
 import { useFetchSearch } from './useItemsGrid'
 
@@ -30,8 +31,4 @@ useIntersectionObserver(
   },
   { rootMargin: '1000px' }
 )
-
-// dynamic grids
-const container = ref<HTMLDivElement | null>(null)
-const { gridCols } = useDynamicGrid(container)
 </script>
